[FIX #13] Gradle Build System Now Works With Scala
