+++ .idea/modules/morganey-intellij-plugin_test.iml
@@ -15,15 +15,24 @@
     <orderEntry type="module-library">
       <library>
         <CLASSES>
+          <root url="jar://$MODULE_DIR$/../../scala/build/lib/morganey.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
+    <orderEntry type="module-library">
+      <library>
+        <CLASSES>
           <root url="jar://$MODULE_DIR$/../../build/lib/morganey.jar!/" />
         </CLASSES>
         <JAVADOC />
         <SOURCES />
       </library>
     </orderEntry>
+    <orderEntry type="module" module-name="scala_main" />
     <orderEntry type="library" name="Gradle: org.jetbrains.kotlin:kotlin-stdlib:1.0.4" level="project" />
     <orderEntry type="library" name="Gradle: org.scala-lang:scala-compiler:2.11.1" level="project" />
-    <orderEntry type="library" name="Gradle: org.scala-lang:scala-library:2.11.1" level="project" />
     <orderEntry type="module-library" scope="RUNTIME">
       <library>
         <CLASSES>
@@ -33,6 +42,7 @@
         <SOURCES />
       </library>
     </orderEntry>
+    <orderEntry type="library" name="Gradle: org.scala-lang:scala-library:2.11.1" level="project" />
     <orderEntry type="module-library">
       <library name="Gradle: com.jetbrains:idea:LATEST-EAP-SNAPSHOT">
         <CLASSES>
