+++ .idea/modules/scala/scala.iml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module external.linked.project.id=":scala" external.linked.project.path="$MODULE_DIR$/../../../scala" external.root.project.path="$MODULE_DIR$/../../.." external.system.id="GRADLE" external.system.module.group="me.rexim.morganey.plugin.scalaobjects" external.system.module.version="0.1" type="JAVA_MODULE" version="4">
+  <component name="NewModuleRootManager" inherit-compiler-output="true">
+    <exclude-output />
+    <content url="file://$MODULE_DIR$/../../../scala">
+      <excludeFolder url="file://$MODULE_DIR$/../../../scala/.gradle" />
+      <excludeFolder url="file://$MODULE_DIR$/../../../scala/build" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+  </component>
+</module>
