+++ build.gradle
@@ -55,7 +55,7 @@ repositories {
 }
 
 dependencies {
-    testCompile 'junit:junit:4.12'
+    compile project(':scala')
     compile urlFile('https://github.com/morganey-lang/morganey-intellij-plugin/raw/new-build/jars/morganey-01.10.2016.1900.jar', 'morganey')
     compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     compile "org.scala-lang:scala-compiler:2.11.1"
