+++ settings.gradle
@@ -1,1 +1,3 @@
 rootProject.name = 'morganey-intellij-plugin'
+include 'scala'
+
